Moved UBC event handler and facebook handler into one file
